fixed messaging

+++ index.js
@@ -17,44 +17,53 @@ const io = socketIO(app, {
 });
 
 io.sockets.on("connection", function (socket) {
-    async function clientLog(message) {
+    async function log(message, sendToClient = false) {
         const arrayLog = [new Date(), socket.id, "Message from server: ", message],
             stringLog = arrayLog.join(" | ");
-        console.log(stringLog);
-        socket.emit("log", stringLog);
+            
+        console.log("log: ", stringLog);
+        if (sendToClient) socket.emit("log", arrayLog);
         try {
-            const stats = await fs.statSync("./logs/logs.txt");
-            if(stats.isFile() && stats.size > 1024000) {//10 MB
+            let stats,
+                dir = "./logs";
+
+            if (!fs.existsSync(dir)){
+                fs.mkdirSync(dir);
                 fs.writeFile("./logs/logs.txt", stringLog, err => {
-                    console.error(err);
+                    if (err) console.error("error: ", err);
                 });
-            } else if(stats.isFile()) {
-                fs.writeFile("./logs/logs.txt", "\n" + stringLog, { flag: 'a+' }, err => {
-                    console.error(err);
-                });
+            } else {
+
+                stats = await fs.statSync("./logs/logs.txt");
+            
+                if(stats.isFile() && stats.size > 1024000) {
+                    fs.writeFile("./logs/logs.txt", stringLog, err => {
+                        if (err) console.error("error: ", err);
+                    });
+                } else if(stats.isFile()) {
+                    fs.writeFile("./logs/logs.txt", "\n" + stringLog, { flag: 'a+' }, err => {
+                        if (err) console.error("error: ", err);
+                    });
+                }
             }
         } catch (err) {
             if(err.code === 'ENOENT') {
                 fs.writeFile("./logs/logs.txt", stringLog, err => {
-                    console.error(err);
+                    if(err) console.error("error: ", err);
                 });
             } else {
-                console.log(err);
+                console.error(err);
             }
         }
     }
 
     function messageProcessing(message) {
-        //console.log("message processing ", message);
-        //console.log(socket);
         if (socket.disconnected === true) {
-            //console.log("remove rooms info for id: ", socket.id);
-            //console.log(socket.adapter.rooms);
+            log("remove rooms info for id: ", socket.id);
+            log(socket.adapter.rooms);
             socket.adapter.rooms.forEach((roomInfo, room) => {
-                //console.log("room: ", room);
-                //console.log(roomsInfo);
                 if (roomsInfo[room]) {
-                    console.log(
+                    log(
                         "disconnect message processing, broadcast to ",
                         socket.id
                     );
@@ -63,15 +72,14 @@ io.sockets.on("connection", function (socket) {
                         .emit("removed", roomsInfo[room][socket.id]);
                     socket.leave(room);
                     delete roomsInfo[room][socket.id];
-                    console.log("disconnected");
-                    console.log(roomsInfo[room]);
+                    log("disconnected");
+                    log(roomsInfo[room]);
                 }
             });
         }
         socket.rooms.forEach((room) => {
-            console.log("room: ", room);
             if (roomsInfo[room] && message === "disconnect") {
-                console.log(
+                log(
                     "disconnect message processing, broadcast to ",
                     socket.id
                 );
@@ -80,12 +88,12 @@ io.sockets.on("connection", function (socket) {
                     .emit("removed", roomsInfo[room][socket.id]);
                 socket.leave(room);
                 delete roomsInfo[room][socket.id];
-                console.log("disconnected");
-                console.log(roomsInfo[room]);
+                log("disconnected");
+                log(roomsInfo[room]);
             } else if (roomsInfo[room] && isCurrentSocketInRoom(room)) {
                 roomsInfo[room][socket.id] = message;
-                console.log("updated info:");
-                console.log(roomsInfo);
+                log("updated info:");
+                log(JSON.stringify(roomsInfo));
                 socket.broadcast.to(room).emit("message", message);
             }
         });
@@ -95,27 +103,26 @@ io.sockets.on("connection", function (socket) {
         return Object.prototype.hasOwnProperty.call(roomsInfo[room], socket.id);
     }
 
-    
-
     socket.on("message", function (message) {
-        clientLog("Client said: ", message);
-        //broadcast only for belong rooms
+        if(message.type !== "move") {
+            log("Message from client: " + JSON.stringify(message));
+        }
         messageProcessing(message);
     });
 
     socket.on("gatherRoomsInfo", () => {
-        console.log("gather rooms info received");
+        log("gather rooms info received");
         socket.emit("roomsInfo", roomsInfo);
     });
 
     socket.on("create or join", function (room, map, maxPlayers = 2) {
-        clientLog("Received request to create or join room " + room);
+        log("Received request to create or join room " + room);
         if (roomsInfo[room]) {
             //@todo: rebuild this stuff
             Object.keys(roomsInfo[room]).forEach((joined_id) => {
                 if (socket.id !== joined_id) {
-                    console.log("sent message, with current info");
-                    console.log(roomsInfo[room][joined_id]);
+                    log("sent message, with current info");
+                    log(roomsInfo[room][joined_id]);
                     io.to(socket.id).emit(
                         "message",
                         roomsInfo[room][joined_id]
@@ -124,34 +131,34 @@ io.sockets.on("connection", function (socket) {
             });
 
             map = roomsInfo[room].map;
-            console.log("map: ", map);
+            log("map: ", map);
         } else {
             roomsInfo[room] = { map };
         }
         roomsInfo[room][socket.id] = {};
         ///////////
-        console.log("client added");
-        console.log(roomsInfo);
+        log("client added");
+        log(JSON.stringify(roomsInfo));
         const clientsInRoom = io.sockets.adapter.rooms.get(room);
         const numClients = clientsInRoom ? clientsInRoom.size : 0;
-        clientLog("Room " + room + " now has " + numClients + " client(s)");
+        log("Room " + room + " now has " + numClients + " client(s)");
 
         if (numClients === 0) {
             socket.join(room);
-            clientLog("Client ID " + socket.id + " created room " + room);
+            log("Client ID " + socket.id + " created room " + room);
             socket.emit("created", room, map);
         } else if (numClients < maxPlayers) {
             socket.join(room);
-            clientLog("Client ID " + socket.id + " joined room " + room);
+            log("Client ID " + socket.id + " joined room " + room);
             socket.emit("joined", room, map);
             io.sockets.in(room).emit("joined");
         } else {
-            // max two clients
             socket.emit("full", room);
         }
+        log("current rooms:", socket.rooms);
     });
 
     socket.on("disconnect", function (reason) {
-        clientLog("received bye");
+        log("received bye");
     });
 });
